Adding Policies

+++ Restaurants/Restaurants.Application/Restaurants/Commands/DeleteRestaurant/DeleteRestaurantCommandHandler.cs
@@ -3,10 +3,16 @@ using Microsoft.Extensions.Logging;
 using Restaurants.Domain.Repositories;
 using Restaurants.Domain.Exceptions;
 using Restaurants.Domain.Entities;
+using Restaurants.Domain.Interfaces;
+using Restaurants.Domain.Constants;
 
 namespace Restaurants.Application.Restaurants.Commands.DeleteRestaurant;
 
-public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandler> logger, IRestaurantRepository repository) : IRequestHandler<DeleteRestaurantCommand, bool>
+public class DeleteRestaurantCommandHandler(
+    ILogger<DeleteRestaurantCommandHandler> logger, 
+    IRestaurantRepository repository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) 
+    : IRequestHandler<DeleteRestaurantCommand, bool>
 {
     public async Task<bool> Handle(DeleteRestaurantCommand request, CancellationToken cancellationToken)
     {
@@ -17,6 +23,9 @@ public class DeleteRestaurantCommandHandler(ILogger<DeleteRestaurantCommandHandl
         {
             throw new NotFoundException(nameof(Restaurant), request.Id.ToString());
         }
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, RessourceOperation.Delete))
+            throw new ForbidException();
 
         await repository.Delete(restaurant);
